backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
pydantic==2.6.1
google-cloud-firestore==2.14.0
google-genai==0.4.0
firebase-admin==6.4.0
python-dotenv==1.0.1
